[Homepage] consolidate ColumnLayout components
