+++ src/homepage/index.js
@@ -2,8 +2,7 @@ import React from 'react'
 import './ColumnLayout.css'
 import './Home.css'
 import { AnnouncementBanner } from './AnnouncementBanner'
-import { Column } from './Column'
-import { ColumnLayout } from './ColumnLayout.jsx'
+import { ColumnLayout, Column } from './ColumnLayout.jsx'
 import { DataBrowser } from './ForDataUsers/DataBrowser'
 import { DataPublication } from './ForDataUsers/DataPublication'
 import { Filing } from './ForFilers/Filing'
